feat(calendar-filters): add filters on calendar page
