+++ app/modules/booking/service.server.ts
@@ -1946,7 +1946,14 @@ export async function getBookingsForCalendar(params: {
     canSeeAllBookings,
     canSeeAllCustody,
   } = params;
-  const searchParams = getCurrentSearchParams(request);
+
+  const { searchParams, search, status, teamMemberIds, tags } =
+    await getBookingsFilterData({
+      request,
+      canSeeAllBookings,
+      organizationId,
+      userId,
+    });
 
   const start = searchParams.get("start") as string;
   const end = searchParams.get("end") as string;
@@ -1956,13 +1963,16 @@ export async function getBookingsForCalendar(params: {
       organizationId,
       page: 1,
       perPage: 1000,
+      search,
       userId,
+      ...(status && {
+        // If status is in the params, we filter based on it
+        statuses: [status],
+      }),
       bookingFrom: new Date(start),
       bookingTo: new Date(end),
-      ...(!canSeeAllBookings && {
-        // If the user is self service, we only show bookings that belong to that user)
-        custodianUserId: userId,
-      }),
+      custodianTeamMemberIds: teamMemberIds,
+      tags,
       extraInclude: {
         custodianTeamMember: true,
         custodianUser: true,
